requirements-dev.txt
pip-tools
pytest-django
pytest-socket
factory-boy